Move to .gltf models, tweak the mill models a bit
Raylib has a problem with the animations from Blender-exported glTF
models where "STEP" in animations breaks the animation playback. This
can be fixed by manually turning STEP to LINEAR in .gltf files, but
Blender will keep changing it back in import/export cycles so it's
simplest to just keep models in .gltf and manually fixing them as
needed.